requirements.txt
pandas
np
plotly
matplotlib
seaborn
folium
streamlit_folium
streamlit_extras
pickle
json